feat: Schema 1.0.0 버전 Migration 완료(#116)
