+++ Projects/Core/Sources/Models/RefactoredLineup.swift
@@ -9,27 +9,27 @@
 import Foundation
 import SwiftData
 
-@Model
-public final class RefactoredLineup {
-    public let id: UUID
-    public var lineupName: String
-    public var uniform: Uniform
-    public var formation: Formation
-    public var primaryColor: UniformColor
-    public var secondaryColor: UniformColor
-
-    public init(
-        id: UUID,
-        lineupName: String,
-        uniform: Uniform,
-        formation: Formation,
-        primaryColor: UniformColor,
-        secondaryColor: UniformColor) {
-            self.id = id
-            self.lineupName = lineupName
-            self.uniform = uniform
-            self.formation = formation
-            self.primaryColor = primaryColor
-            self.secondaryColor = secondaryColor
-        }
-}
+//@Model
+//public final class RefactoredLineup {
+//    public let id: UUID
+//    public var lineupName: String
+//    public var uniform: Uniform
+//    public var formation: Formation
+//    public var primaryColor: UniformColor
+//    public var secondaryColor: UniformColor
+//
+//    public init(
+//        id: UUID,
+//        lineupName: String,
+//        uniform: Uniform,
+//        formation: Formation,
+//        primaryColor: UniformColor,
+//        secondaryColor: UniformColor) {
+//            self.id = id
+//            self.lineupName = lineupName
+//            self.uniform = uniform
+//            self.formation = formation
+//            self.primaryColor = primaryColor
+//            self.secondaryColor = secondaryColor
+//        }
+//}
