+++ Projects/Core/Sources/Models/RefactoredTeam.swift
@@ -8,31 +8,31 @@
 
 import Foundation
 import SwiftData
-
-@Model
-public final class RefactoredTeam {
-    public let id: UUID
-    public var teamName: String
-    public var subTitle: String
-    public var isSelected: Bool
-    public var lineups: [RefactoredLineup]
-    public let createdAt: Date
-    public var updatedAt: Date
-
-    public init(
-        id: UUID,
-        teamName: String,
-        subTitle: String,
-        isSelected: Bool,
-        lineups: [RefactoredLineup],
-        createdAt: Date,
-        updatedAt: Date) {
-            self.id = id
-            self.teamName = teamName
-            self.subTitle = subTitle
-            self.isSelected = isSelected
-            self.lineups = lineups
-            self.createdAt = createdAt
-            self.updatedAt = updatedAt
-        }
-}
+//
+//@Model
+//public final class RefactoredTeam {
+//    public let id: UUID
+//    public var teamName: String
+//    public var subTitle: String
+//    public var isSelected: Bool
+//    public var lineups: [RefactoredLineup]
+//    public let createdAt: Date
+//    public var updatedAt: Date
+//
+//    public init(
+//        id: UUID,
+//        teamName: String,
+//        subTitle: String,
+//        isSelected: Bool,
+//        lineups: [RefactoredLineup],
+//        createdAt: Date,
+//        updatedAt: Date) {
+//            self.id = id
+//            self.teamName = teamName
+//            self.subTitle = subTitle
+//            self.isSelected = isSelected
+//            self.lineups = lineups
+//            self.createdAt = createdAt
+//            self.updatedAt = updatedAt
+//        }
+//}
