+++ Projects/Core/Sources/Models/Team.swift
@@ -10,29 +10,29 @@ import Foundation
 import SwiftUI
 import SwiftData
 
-@Model
-public final class Team {
-    public let id: UUID
-    public var teamName: String
-    public var isSelected: Bool
-    public let createdAt: Date
-    public var updatedAt: Date
-    public var lineup: [Lineup]
-    public var teamMembers: [Human]
-
-    public init(id: UUID,
-                teamName: String,
-                isSelected: Bool,
-                createdAt: Date,
-                updatedAt: Date,
-                lineup: [Lineup],
-                addedPlayers: [Human]) {
-        self.id = id
-        self.teamName = teamName
-        self.isSelected = isSelected
-        self.createdAt = createdAt
-        self.updatedAt = updatedAt
-        self.lineup = lineup
-        self.teamMembers = addedPlayers
-    }
-}
+//@Model
+//public final class Team {
+//    public let id: UUID
+//    public var teamName: String
+//    public var isSelected: Bool
+//    public let createdAt: Date
+//    public var updatedAt: Date
+//    public var lineup: [Lineup]
+//    public var teamMembers: [Human]
+//
+//    public init(id: UUID,
+//                teamName: String,
+//                isSelected: Bool,
+//                createdAt: Date,
+//                updatedAt: Date,
+//                lineup: [Lineup],
+//                addedPlayers: [Human]) {
+//        self.id = id
+//        self.teamName = teamName
+//        self.isSelected = isSelected
+//        self.createdAt = createdAt
+//        self.updatedAt = updatedAt
+//        self.lineup = lineup
+//        self.teamMembers = addedPlayers
+//    }
+//}
