+++ Projects/Core/Sources/Models/Theme.swift
@@ -8,14 +8,14 @@
 
 import Foundation
 
-public enum Theme: Codable {
-    case blueGray
-    case whiteGreen
-    case blackBlue
-    case grayBlack
-    case blueBlack
-    case blackGray
-    case purpleClear
-    case blueClear
-    case neonPurple
-}
+//public enum Theme: Codable {
+//    case blueGray
+//    case whiteGreen
+//    case blackBlue
+//    case grayBlack
+//    case blueBlack
+//    case blackGray
+//    case purpleClear
+//    case blueClear
+//    case neonPurple
+//}
