+++ Projects/Feature/Sources/Observable/Edit/ThemeObservable.swift
@@ -17,11 +17,11 @@ final class ThemeObservable {
         .whiteGreen,
         .blackBlue,
         .grayBlack,
-        .blueBlack,
-        .blackGray,
-        .purpleClear,
-        .blueClear,
-        .neonPurple
+//        .blueBlack,
+//        .blackGray,
+//        .purpleClear,
+//        .blueClear,
+//        .neonPurple
     ]
     var lineup: Lineup
