+++ Projects/Feature/Sources/View/Temp/LineupCRUDView.swift
@@ -9,42 +9,42 @@
 import Core
 import SwiftData
 import SwiftUI
-
-public struct LineupCRUDView: View {
-    var team: RefactoredTeam
-
-    @Environment(\.modelContext) var context
-
-    public var body: some View {
-        VStack {
-            List {
-                ForEach(team.lineups, id: \.self) { lineup in
-                    Text(lineup.lineupName)
-                }
-                .onDelete(perform: deleteLineup)
-            }
-
-            HStack {
-                Button(action: {
-                }, label: {
-                    Text(String(localized: "Add Lineup"))
-                })
-                .padding(.top, 10)
-            }
-        }
-    }
-
-    func deleteLineup(at offsets: IndexSet) {
-        do {
-            for index in offsets {
-                let teamToDelete = team.lineups[index]
-                //            context.delete(teamToDelete)
-            }
-            team.lineups.remove(atOffsets: offsets)
-
-            try context.save()
-        } catch {
-            print("error")
-        }
-    }
-}
+//
+//public struct LineupCRUDView: View {
+//    var team: RefactoredTeam
+//
+//    @Environment(\.modelContext) var context
+//
+//    public var body: some View {
+//        VStack {
+//            List {
+//                ForEach(team.lineups, id: \.self) { lineup in
+//                    Text(lineup.lineupName)
+//                }
+//                .onDelete(perform: deleteLineup)
+//            }
+//
+//            HStack {
+//                Button(action: {
+//                }, label: {
+//                    Text(String(localized: "Add Lineup"))
+//                })
+//                .padding(.top, 10)
+//            }
+//        }
+//    }
+//
+//    func deleteLineup(at offsets: IndexSet) {
+//        do {
+//            for index in offsets {
+//                let teamToDelete = team.lineups[index]
+//                //            context.delete(teamToDelete)
+//            }
+//            team.lineups.remove(atOffsets: offsets)
+//
+//            try context.save()
+//        } catch {
+//            print("error")
+//        }
+//    }
+//}
